Directly pass timer configuration to initializers
Instead of using a timer_configuration struct, all settings are now
specified directly in the call to timer<n>_init. Default values are now
available as #defines.

In addition to this conversion, the input capture noise canceler setting
is now an enum instead of a bool, to make its meaning more explicit in
user code.

Lastly, a little more documentation was added and the static variable
wgm_error was renamed to wgm_mode_invalid;